Fix import model dtype bug

+++ gr00t/model/policy.py
@@ -229,7 +229,7 @@ class Gr00tPolicy(BasePolicy):
         return True
 
     def _load_model(self, model_path):
-        model = GR00T_N1.from_pretrained(model_path)
+        model = GR00T_N1.from_pretrained(model_path, torch_dtype=torch.bfloat16)
         model.eval()  # Set model to eval mode
         model.to(device=self.device)  # type: ignore
